Cache political bias results based on md5 hash

+++ project_files/ai_analysis/tasks.py
@@ -1,41 +1,54 @@
-from scraper.utils import scrape_data
-from .utils import analyse_sentences_for_bias, extract_ideal_sentences, text_to_md5_hash
+from .utils import analyse_sentences_for_bias, extract_ideal_sentences
 from celery import shared_task
 from .models import ArticleAnalysisResults
 from traceback import format_exc
 from collections import Counter
+from scraper.utils import scrape_data
+from .utils import text_to_md5_hash
 
 
 # Shared task is basically telling the function to be picked up by a celery worker
 # This lets me hook onto the synchronous scraper function I implemented in scraper.utils
 # with an asynchronous version that calls it but does not interfere with its functionality or unit tests
 @shared_task
-def scrape(url, analysis_id):
+def async_scrape(url, analysis_id):
     try:
         # Scrape the article content
         article_text = scrape_data(url)
+
+        # continue to analyze the scraped data
+        begin_analysis.delay(analysis_id, article_text)
     except Exception as e:
         # Log an error message and mark the analysis as failed
         print(
             f"An error occurred when scraping from {url}: {e}\n{format_exc()}")
         analysis_failed(analysis_id)
 
-    try:
-        # Extract sentences for analysis
-        sentences_df = extract_ideal_sentences(article_text)
 
+@shared_task
+def begin_analysis(analysis_id, article_text):
+    try:
         # Calculate the MD5 hash of the article text
         article_text_md5 = text_to_md5_hash(article_text)
 
-        # Convert the sentence dataframe to json so Celery can serialize it
-        sentences_df = sentences_df.to_json(orient='split')
+        # Check for existing analysis with the same MD5 hash
+        existing = ArticleAnalysisResults.objects.filter(
+            article_text_md5=article_text_md5).first()
 
-        # Pass the MD5 hash along with other data to the analysis task
-        analyse_sentences.delay(analysis_id, sentences_df, article_text_md5)
+        if existing:
+            # If existing analysis is found, update the current analysis entry with its details
+            update_analysis_with_existing(analysis_id, existing)
+        else:
+            # If no existing analysis is found, start the analysis process
+            # Extract sentences for analysis
+            sentences_df = extract_ideal_sentences(article_text)
+            # Convert the sentence dataframe to json so Celery can serialize it
+            sentences_df = sentences_df.to_json(orient='split')
+            analyse_sentences.delay(
+                analysis_id, sentences_df, article_text_md5)
     except Exception as e:
         # Log an error message and mark the analysis as failed
-        print(
-            f"An error occurred when extracting ideal sentences: {e}\n{format_exc()}")
+        print(f"An error occurred during processing: {e}\n{format_exc()}")
         analysis_failed(analysis_id)
 
 
@@ -47,8 +60,7 @@ def analyse_sentences(self, analysis_id, sentences_df, article_text_md5):
         bias_results = analyse_sentences_for_bias(sentences_df)
 
         # Store the analysis results
-        analysis_complete(analysis_id, bias_results,
-                          article_text_md5)
+        analysis_complete(analysis_id, bias_results, article_text_md5)
 
     except ValueError as exc:
         # If an error occurs, retry the task up to 3 times
@@ -89,9 +101,21 @@ def analysis_complete(analysis_id, bias_results, article_text_md5):
     # Update the analysis object in the database with the results
     analysis = ArticleAnalysisResults.objects.get(id=analysis_id)
     analysis.url = analysis.url
-    analysis.article_text_md5 = article_text_md5
     analysis.sentence_results = bias_results
+    analysis.article_text_md5 = article_text_md5
     analysis.bias_conclusion, analysis.bias_percent = make_final_conclusion(
         bias_results)
     analysis.status = 'completed'
+    analysis.save()
+
+
+def update_analysis_with_existing(analysis_id, existing):
+    # Copy the results from the already complete analysis
+    analysis = ArticleAnalysisResults.objects.get(id=analysis_id)
+    analysis.url = existing.url
+    analysis.sentence_results = existing.sentence_results
+    analysis.article_text_md5 = existing.article_text_md5
+    analysis.bias_conclusion = existing.bias_conclusion
+    analysis.bias_percent = existing.bias_percent
+    analysis.status = existing.status
     analysis.save()
